refactor: merge constants.py into config.py
Phase A. Five module-level values sat in their own module with nothing separating them from
the ~150 lines in config.py — no rule said which file a new constant belonged in, so the
answer was whichever one the author happened to open.

They join SEARCH_CAPTURE_TIMEZONE at config.py's module level rather than being folded into
the Config dataclass. Config's attributes are evaluated at class-definition time and several
read QgsSettings; these five are literals and have no reason to acquire that.

model/provider now imports config where it imported constants. That direction is allowed —
spec/007_architecture.md § Layer rules lets model/ read config — and it is the only new edge:
config imports nothing from the plugin, so it stays a leaf.

## Manual test

Regression surface — the moved values are each read on a path worth one click:
- add a user XYZ provider, restart QGIS, confirm it is still listed (PROVIDERS_KEY)
- preview a Mapbox source, which cannot preview, and confirm the OpenStreetMap layer is added
  instead (OSM)
- confirm the source combo still offers "Imagery Search" and "My imagery" by those names
- any request completing normally exercises the default timeout

New behaviour: none.
