--- mapflow/constants.py
@@ -1,8 +1,0 @@
-PROVIDERS_KEY = 'mapflow_data_providers'
-
-OSM = 'type=xyz&url=https://tile.openstreetmap.org/{z}/{x}/{y}.png&zmax=19&zmin=0'
-
-SEARCH_OPTION_NAME = "🔎 Imagery Search"
-CATALOG_OPTION_NAME = "🖼️ My imagery"
-
-DEFAULT_HTTP_TIMEOUT_SECONDS = 10
